moved array utils to its own module
